% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IGAP論文.R
\name{iGAP}
\alias{iGAP}
\title{Symbolic interval data with Generalized Association Plots}
\usage{
function(data = NULL,
         standardize_condition = TRUE,
         cor_up = "centers",
         distance_right = "Hausdorff",
         normalize_right = TRUE,
         SpanNormalize_right = TRUE,
         euclidea_right = TRUE,
         method_up = NULL,
         method_right = NULL,
         Rcolor = "Spectral",
         Rcolorr = "Spectral",
         Rcoloru = "RdBu")
}
\arguments{
\item{data}{A iGAP object. It can also be either RSDA object or
classical data frame.}

\item{standardize_condition}{Boolean variables, which mean the color
present by standardize condition (if TRUE) or matrix condition (if FALSE).}

\item{cor_up}{For the upper-side association correlation matrix using
methods (centers, BD, BG)}

\item{distance_right}{For the right side association distance matrix using
methods (Gowda.Diday, Ichino, Minkowski, Hausdorff)}

\item{normalize_right}{A logical value indicating whether normalize the data in the ichino or hausdorff method.}

\item{SpanNormalize_right}{A logical value indicating whether.}

\item{euclidea_right}{A logical value indicating whether use the euclidean distance.}

\item{method_up}{For the upper-side association distance matrix use sorting
methods (R2E, GW, HC, OLO, MDS, isoMDS, isomap, monoMDS, metaMDS, Sammon,
ARSA, GSA, SGD, QAP, TSP, Identity, Random,  Reverse, SPIN, VAT, ward.D,
ward.D2, single, complete", average , mcquitty , median, centroid)}

\item{method_right}{For the right side association distance matrix use sorting
methods (R2E, GW, HC, OLO, MDS, isoMDS, isomap, monoMDS, metaMDS, Sammon,
ARSA, GSA, SGD, QAP, TSP, Identity, Random, Reverse, SPIN, VAT, ward.D,
 ward.D2, single, complete", average , mcquitty , median, centroid)}

\item{Rcolor}{Color scale to use for interval data.}

\item{Rcolorr}{Color scale to use for the right side association distance matrix.}

\item{Rcoloru}{Color scale to use for the upper-side association correlation matrix.}
}
\value{
Return a ggplot2 object.
}
\description{
Let interval data be visualized as a heatmap, and the correlations heatmap
between interval data variables and the distances heatmap between concepts can be collaged to
explore the hidden structure of interval data.
}
\examples{
p <- iGAP(facedata)
p

test <- classic.to.sym(x = ex1_db2so, concept = c(state, sex), variables = c(county, group, age))
p <- iGAP(test, standardize_condition = F,
          cor_up = c("BG"),
          distance_right = "Hausdorff",
          normalize_right = T,
          SpanNormalize_right = T,
          euclidea_right = T,
          method_up = "R2E",
          method_right = "R2E",
          Rcolor = "Spectral",
          Rcolorr = "Spectral",
          Rcoloru = "RdBu")
p

p <- iGAP(test, standardize_condition = F,
          cor_up = c("BG"),
          distance_right = "Hausdorff",
          normalize_right = T,
          SpanNormalize_right = T,
          euclidea_right = T,
          method_up = "single",
          method_right = "single",
          Rcolor = "Spectral",
          Rcolorr = "Spectral",
          Rcoloru = "RdBu")
p

}
